Make Hoarfrost Cannonade monster-castable
To keep monster from wasting their cannon charge-up, they will only recast
the spell if no existing cannon of theirs is in sight of their foe.

(Getting the cannons to disappear on the death of their creator involved
some fairly ugly hacks, since the main method of doing this explicitly only
works on things which are abjurable. In the future, it would be nice to
unify the various non-abjurable things that go away on creator death (ie:
tentacles, battlespheres, vines, spectral weapons, and now this) instead of
each one being handled essentially in its own one-off way.

This commit modifies monster::is_summoned to retrieve the summon type of
a creature regardless of whether they're abjurable or not (since many
non-abjurable things set this!). It was unclear to me why it would only do
this for abjurable things. Since the return value of the method is
unaffected, I don't THINK this changes the semantics of anything currently
using it.