Replace Deep Elf Elementalist Freezing Cloud with Hoarfrost Cannonade
Freezing cloud wasn't a very functional spell on these enemies. Because of
the unpredictable splash of monster cloud spells, the elementalist would
often refuse to cast it at all (even in an open room with no one else
around!) for fear of hitting themselves. And then when they did, it was
often more dangerous to the elementalist (or other elves) than it was to
the player. Let's try replacing it with a different ice spell that has
neither of these problems.

I am letting monsters 'cheat' and create the cannons centered around their
foe, rather than themselves (both so that they can do something with it
when there's a pileup of other elves in the way, and so it's less trivial
for the player to just stay out of range of them).