Tweak Cloak of the Thief
While noise suppression is a powerful effect, many players felt
uncomfortable using this item now that it had an even greater slaying
penalty than the old version.

But it occurred to me that Quiet is already less useful on the sorts of
characters most affected by -slay, so let's try making it give a moderate
MP reduction instead (since Quiet is strongest on casters).