Blindness: player accuracy debuff
Monster blindness has been available in various forms for some time, but
there was no equivalent player effect and obvious sources of blindness just
Confused the player if anything.

The new Blind status can be inflicted on the player with a wand of light
and with atropa darts.

While affected, your accuracy is reduced; by more the further away the
target is. Mechanically your "to hit" roll is reduced by 10% for every tile
distance (starting at your own tile) so maximum 80% for a normal race or
90% for Barachi. It affects melee and ranged, and also those spells that
care about accuracy, so this has most effect on ranged combat.