Tune down quicksilver bolt by 3d5 to bring more in line with other breaths
Quicksilver bolt is powerful, accurate, irresistible, and has a strong effect.
The damage of quicksilver matches other breaths which don't have a
special effect. To bring it more in line with acid breath, which is tuned to a
lower damage level to account for the corrosion, reduce the damage by 3d5,
bringing purple draconian from 3d19 to 3d14, and quicksilver dragon from
3d20 to 3d15.