Fix a Hellfire Mortar crash
It is possible that the target of the second bolt of magma in a turn will
be dead by the time it comes to fire on it (probably because it was also
in line of the first bolt, even if not its primary target). Instead of
assuming each target we picked is alive, double-check and then skip to the
next candidate, if the original is dead and one exists. That way the mortar
won't sometimes appear to mysteriously only fire once, even when a viable
second shot clearly exists.