Fix missing break statements in spell usability logic (RelicWraith)
Due to an oversight that was then repeatedly several times in a row, a
number spells were checking the usability of both themselves *and* every
spell implemented after that.

I'm impressed this didn't come up sooner, but I guess a lot of these spells
are unlikely to be used at the same time by the same character.