Make the orc priest apostles a little more fancy (#3628)
To differentiate orc priest apotles more from just being recoloured orc
wizards holding the orc priest tablet, give them a bit of shiny trim like
high priests have, but with silvery-metal colours and skipping the collar
to reduce the chance of it being confused with said high priests (who have
gold trim and trim on the collar). Ideally in the future we could get
fancier and slightly varying tiles on this front, but this should help for
now.

Closes #3628.