Implement backbone of new spell school: Forgecraft
Forgecraft is about making tangible physical constructs and reinforcing
them in various ways. It contains many spells which create things like
cannons, golems, or magical weapons, as well as a variety of supportive
utility.

This school is not intended to address a specific 'flaw' in Crawl at the
moment, but rather because I think it would be a fun new thing and allow
for even more cross-school diversity. I believe there is still a lot of
untapped potential in ally-creation spells, but Summonings was already one
of the most populated schools in the game. Necromancy shows that other
schools with an ally+utility focus can work well in Crawl, and this is
intended to be a new one of such - landing somewhere between the two in
terms of ally/utility balance.

Compared to Summonings, Forgecraft allies tend to have more specific unique
behaviors, and either operate in a limited way or are triggered based on
the caster's own actions (ie: as Battlesphere is). Forgecraft is also aimed
with hybrid play in mind from the start - with multiple spells that benefit
from melee cooperation in specific ways that Summoning spells don't or
which are directly augmented by melee ability.

Multiple existing spells will be moved into Forgecraft (with some
adjustments), as well as many new spells across subsequent commits.

This commit adds the basic infrastructure of the Forgecraft skill, as well
as icons, descriptions, Forgecraft enhancer artprop, Ru sac skill, randbook
words, and so forth, as well as a placeholder for a new mage background,
Forgewright.

(Because of strict skill enum grouping among spell skills (ie: all of them
have to be contiguous), this repurposes the SK_TRANSMUTATIONS enum to
become SK_FORGECRAFT. But, after examining this, I think it should be
basically safe.)

Aptitudes:

Mountain Dwarf: +2 (but now have -2 Summoning)
Coglin: +2
Yellow Draconian: +2 (they were the only colour with no aptitude bonus,
                      and this seemed reasonably connected, since Alchemy
                      was already used.)
Demonspawn: -1
Felid: -1
Ghoul: -2
Minotaur: -2
Tengu: -2 (to contrast with their +2 summoning, as an inverse of MD)

In all other cases, species have the same Forgecraft apt as they had
Summonings.

Forgecraft miscasts cause self-corrosion - less because acid damage is a
theme in Forgecraft spells, and more that operating upon the structural
integrity of physical objects is their domain, and making said structure
*worse* feels like a plausible failure state to performing them
incorrectly.