Revised formula for player blindness
The old formula produced weird cliffs where chance dropped rapdily to 2% at
an arbitrary distance, because we were scaling to_hit to *below* the
target's ev where it is no longer possible to hit them.

The new formula treats target ev as a "minimum" to-hit and scales to near
that point instead, using a simple inverse formula to produce a curve. This
provides a much smoother and more predictable change over distance while
still meeting the basic requirements of "noticable in melee, but extreme at
LOS range". The malus is move to pre_roll rather than post_roll as this
adjustment *has* to be made before the roll when we still know the margin.

Another surprise: it escaped my attention before that ev is applied wildly
different in beam accuracy vs regular combat; for beams, the ev is
randomised so the effective useful to-hit is 0 instead of the targets ev.
So the old formula was already behaving quite reasonably for beams, but the
new formula didn't: so now beams pretend that ev is 0 for the purpose of
the blindness malus.

Other minor changes:
 - Tweaked language to be less ambiguous
 - Removed blindness preventing you read scrolls
 - Don't specifically wake player up on blinding (the cause of the blinding
   will probably wake them anyway)
 - Slightly shuffle messaging and other side effects so they don't trigger
   if blindness didn't increase
 - Cleaned up an unused key and some unneccessary consts
 - Fixed some Lua interface comments
 - Split out the description function from the calculation so the same
   rendering can be used for beam targetting