Make Magnavolt beams non-bouncy
In most circumstances, this doesn't matter, since the beams are always
aimed at a monster and stop at that monster. But if one of the magnetised
targets ends up on the other side of a granite statue, you can still see
them, but the beam cannot properly reach them, and will rebound off the
statue without a warning.

Rather than deal with all the targeter issues of properly displaying and
prompting the player for these situations (since that would only make the
spell feel cludgier), let's just make the beams stop if they can't reach
their target.

Fixes #3809