Don't perform a deathbed conversion on an already-converted orc
I'm not sure that this deal with #3808 entirely, since I wasn't able to
reproduce the unlinked items part, but in the process of testing, I noticed
that an orc could surrender multiple times if you 'killed' them again after
they'd already converted. This shouldn't happen, regardless, and *maybe*
could be related?