Add a proper targeter for Slouch, display damage per monster
Instead of just hilighting all of LoS when aiming the ability, it will now
actually show which monsters will be affected by Slouch (ie: indicating
that it cannot affect sleeping or stationary monsters) as well as indicate
how much damage an individual monster is expected to take, based on its
speed.

(This should help give players a better understanding of how much speed
affects the damage Slouch actually does)

Also make the ability's damage description more explicit that this is
damage against normal-speed monsters. Unlike spell damage, this is never
listed in a format where we only have a few characters to work with, so
we're free to be a little more verbose.