Swap colours of Shifter and Knight draconian jobs
Purple draconian shifters have been very troublesome to deal with, having a lot
of control on the battlefield being able to blink themselves, the player, and
surround the player with other threats. This comes along with the quicksilver
breath and the according Will increase making the player not be able to deal
with them using a lot of the will checking tools. This change pushes them
more toward the category of berserk moths and doom hounds of a high priority
threat regularly susceptible to a counter instead of being a definite standout
among the other draconian jobs. Also it combos nicely with being able to
reposition the player for double zaps with both of the blinks.