Give Coglins invisibility in Fedhas' Mad Dash.
Instead of a non-draining amulet of invisibility, which they can't wear,
give them a non-draining scarf of invisibility instead. (Adapt code from
meatsprint to accomplish this.)

The item tile is a recolour of item/armour/scarf1.png, using the palette
of item/amulet/artefact/urand_amulet_invisibility.png, and the worn tile
is a recolour of player/cloak/scarf_magenta.png.

Closes #3773.