Fix a crash-prone way of getting more than 3 apostles at once
If you dismissed an apostle who was in a disconnected branch (eg: a portal
vault or the Abyss) that the player was not currently in, it would delete
their apostle data, but *wouldn't* properly remove them from the interlevel
recall list. Officially, they weren't your apostle anymore, but you could
still recall them. And if you did, from there on out, whenever they died,
the game would crash (as it attempted to update their apostle status to
dead, when they no longer had stored apostle data at all).

This should fix that problem.