Experiment with making bands a little harder to separate
Monsters in bands that have a natural leader try to stay within a moderate
distance of this leader while wandering about, so that if the leader is
chasing after the player, the rest of the band tends to follow (whether
they've yet noticed the player themselves or not). But no such protections
between band followers and each other exist, which sometimes results in it
being straightforward to pick off the members of a band individually
without alerting most of them - often without much deliberate effort, if
they're already awake.

In past, there have been some issues with the functionality of bands that
rely on each other for synergy, so I'm curious to try making them a little
bit stickier.

This commit makes it so that when a band leader would pick a new wander
target, if one of their followers is out of sight and seeking a target,
they will wander in the direction of that band member instead of randomly.
And since band followers already vaguely follow the leader, the band as a
whole will tend to move in that direction.

There is still a bit of lag time here. The leader still finishes moving
wherever they were wandering first, and gets no special awareness of the
player (or whatever else their follower was chasing). But it increases the
*tendancy* for the rest of a band to follow on a slight delay from the
player engaging any part of it.

I'm not certain if this change will be impactful enough to be clear to
players, or that it will be a mechanical improvement if it is, but things
of this nature are hard to know without playtesting, so let's playtest!

(This commit also independently fixes a bug introduced by 672eff3212572a203b58ffd4f1e2474d8a40fcfc
which made band leader unable to push past their own followers)