Replace RevMP gizmo property with RevMPSaver
RevMP's effect was widely considered undertuned, but I'd like *some* effect
based on Rev that supports a blaster playstyle, so this is a second
attempt.

RevMPSaver gives a flat cost reduction to the player's spells based on how
revved they are (-1/2/3 MP per rev tier). This effect cannot reduce a spell
below 1 cost.

This has some related dynamics to the old effect, but notably encourages
the player to unload spells *after* melee, which is usually the inverse of
what you'd like to do when blasting. If the power of the effect is strong
enough, I think that could be a distinctive quirk. Wind up a bit to
rapid-fire mid-level conjurations with a hefty discount, or to squeeze the
most value out of the last few MP you have in a prolonged battle.

Let's see how this iteration feels.