Boost Coglin base Will, replace Will+ gizmo property (Lici)
Coglins have often been mentioned as a species that struggles with Will
checks, due to their more restricted access to Will+ items while only
having baseline Will. I'd been mostly okay with this being a weakpoint of
the species, but after some discussion about how this may bias gizmo
selection unduely towards whichever ends up with Will+ on it, I've decided
to experiment with removing it.

There's a fairly limited list of possible bonuses that could replace Will+
in the minor gizmo slot (I feel See Invisible and Flight are too weak to
feel compelling, even paired as a minor bonus, and I'd been deliberately
avoiding AC/EV and stat upgrades), and what I've settled on for now is
Wizardry+4MP. It's possible this doesn't appeals to enough coglins to
open up the decision space between gizmos in the way this commit was
generally intended, but wizardry on an otherwise-appealing gizmo might
specifically incentivize a little hybridization that a player wouldn't have
otherwise done, and that feels worthy trying.