Remove Dithmenos's Shadow Form ability and associated form
Shadow Form was a moderately useful defensive tool, but it came with an
extremely long list of assorted bonuses and penalties (several of which
were highly specific). As part of the pending revamp, it is slated to be
replaced with a different ability that aims to achieve a similar function
of 'powerful defensive cooldown that makes you supernaturally hard to be
seen' in a more streamlined manner. (See 7444593a7a )