Overhaul Dithmenos shadow mimic (for attacks)
Instead of briefly placing an invisible monster at the player's location
(which tries very hard to convince the code in several places that it
really *is* the player - a source of multiple bugs in past), player shadows
are now proper monsters that show up on a nearby tile when they trigger.

(They do not act independently, so they can be somewhat compared to a
spectral weapon that lingers around for a few turns after activating to
potentially distract monsters, and will interact with a future ability.)

Melee attacks that trigger mimic will attempt to place it on the opposite
side of the creature you're attacking and then swing at it, where this is
possible, but can place on any tile up to 3 tiles away, and select any
valid enemy in this range if it is not. (ie: they will always try to hit
something, if they can). Ranged mimics will simply place somewhere nearby,
but otherwise follow similar rules.

The player and their shadow can freely fire through each other.

As before, your shadow wields an unbranded +0 copy of your weapon's base
type, though its own base damage now scales with your XL so that it should
have more impact and remain more useful across the length of the game.

(The logic for spell mimic will be added in a future commit)