Buff / Fix Rimeblight
After 799ce95f6cd29deaeda1c58a8f94fcd328702e82 is was intended that
Rimeblight spread to adjacent creatures whenever an infected creature died,
but in actuality, this only happened if they died due to the instakill
threshold effect and left behind a pillar of rime. As it is extremely
common for a monster to go from >20% hp to dead directly, the buff wasn't
actually working as intended. I have fixed that now.

I suspect this change alone does a very substantial amount to make the
spell feel stronger, but in response to feedback, I have also let
Rimeblight affect nonliving/undead monsters (with the caveat that they
*don't* get the instakill pillar effect, as they have no lifeforce to
empower the plague on the verge of their death), and nudged up the low
end of its duration and DoT damage by a tiny smidge.

Let's see if this finally feels appropriately strong in the proper
situations.