Txt-ify miscast messages
Moving these to their own txt file was reasonably straightforward and
makes it now possible to translate them - as well as making it easier to
make additional substitutions such as colour names or further
randomisations (without having to invoke these lookups unneccesarily for
cases that are not using them).

This commit replicates the exact messages as they were without further
modification.