New Dith 5* ability: Primordial Nightfall
Immediately reduces the player's sight radius to 0 and then slowly restores
it to normal over the duration of the effect. Cannot be reapplied until it
has fully worn off (so the player cannot spam it to remain
nearly-invulnerable).

Fills a little of the role of Shadow Form being a defensive panic button,
without nearly so many different properties and up/downsides baked into it,
and hopefully feels reasonably unique as well.

One quirk that might not be immediately apparent is that enemies can still
attack you when your sight radius is 0 because 'moving into a space' always
counts as an attack if something is there. You do get the invisibility
bonus to avoid their attacks because they still can't 'see' you, however.