Aphotic Marionette (new Dithmenos 4* ability)
An activated ability that allows the player to grab hold of an enemy
through their shadow and force them to unleash a barrage of spells as if
it were the player casting them.

In a single action, the enemy will run through their known spells (in
random order), attempting to cast each on random targets hostile to the
player, and repeating spells only once all valid spells have been cast
once. Creatures they summon this way will belong to the player and many
beneficial effects that they would apply to themselves will be applied to
you instead (such as Might or Invisibility). You are also fully protected
from any damage they would cause (ie: if you force Nikola to repeatedly
cast Chain Lightning).

A handful of spells are not eligible to be cast this way (either because
they would be useless, harmful, or are currently broken in some fashion),
but these are all clearly indicated with an ‘!’ when examining an enemy.

Can only ever be used successfully on given enemy a single time, period.
(In some cases, such as an enemy having a single spell that is
theoretically valid but which no enemy is in range of, using Marionette may
have no useful effect. In those situations, most of the piety cost is
refunded and you will able to try again another turn).

By default, attempts to cast 3 spells by the targeted enemy, though at very
high levels of invocations, this number can be raised as high as 6.

----

The implementation of this is fairly complicated, but I have mostly tried
to avoid doing anything *inside* the code of existing spells (with just a
couple small exceptions). However, wonky as it is in some ways, I think the
net effect feels really cool and meaningfully different than charming, and
rounds out Dith's kit as abilities that never directly damage or summon
anything but manipulate existing monsters and their ability to act against
you through trickery.

(I do plan to do a little more refactoring to this later, but it should
largely function as-is. Somewhat unclear about what the piety cost should
be, since it varies from fairly disappointing to extremely powerful
depending on what is around to use it on, but that's some of what
experimental branches are for)