Move Xom's heel-clicking message to the database.
Simplify things by not having it require boots anymore, and adding it
in the impromptu dance messages, since it could be seen as a dance move.
One side effect of this is that you can't do it while airborne anymore,
but otherwise it has all the same restrictions as Xom's impromptu
dances.