Smooth out stat modifiers for weapon/unarmed damage
Currently, Crawl's uses integer division when applying stat modifiers to a weapon's base damage, which results in sharp breakpoints: using a dagger, the player does exactly the same amount of damage with Dex 29 as they do with Dex 20. This is counterintuitive, misleading, and inconsistent with Crawl's design philosophy.

This pull request implements the simple fix of replacing integer division with random rounding.