Rescale Drain status light colors
The thresholds currently used seem to have been derived directly from the
numbers used when drain affected skill levels instead of HP and are oddly
spaced (changing severity at >0%/5%/10%/25%/50% max HP lost), as well as
(in my opinion) giving a false impression of the severity of the effect at
low levels of draining. (At least I was surprised that red 'heavily
drained' starts at only 10% hp loss, which doesn't seem alarming enough to
warrant that label).

This commit respaces that to >0%/10%/20%/30%/50%. I'm interested in a few
more adjustments to draining formulas, but for now this allows a better
visible threshold for when Midnight Pantomime turns off.