Tweak wanderer good-throwing/good-weapon items
Previously, the good set of throwing items for wanderer was 1-4
curare-tipped darts and a +2 dagger. Neither of these scale well with
throwing skill, which is likely to be your highest skill. This could
also be extremely disappointing if you got 1 curare dart and it
mulched immediately. To encourage actually training throwing skill,
replace the enchanted dagger with large rocks/javelins/boomerangs
(depending on size), up to a total of 7 pieces of ammunition.
(Note: this uses the same code as the gladiator background, so
small species will get double the number of boomerangs.)

Also, the good set of weapon items for all weapon skills had a 4/5
chance of giving a +2 enchantment to a decent-quality weapon
(falchion, mace, etc), and a 1/5 chance of upgrading to a good-quality
weapon (long sword, flail, etc). In practice this upgrade was rather
rare, and starting with a +2 decent weapon is a little too frequent.
So let's bump the upgrade chance to 1/3, for the sake of variety.

[Committer's notes: Edited commit message.]