Try to avoid an infinite handle_monsters loop with Hellfire Mortar
In some circumstances, it was possible for the mortar to be unable to move,
yet not die either, in which case it never consumed energy on its turn and
would cause an infinite handle_monsters() loop.

While some of the individual causes of this warrant being addressed
directly, for now put in general protection so that these should not crash
(but simply have the mortar pass its turn harmlessly and expire at the end
of the lava effect at the latest).