Revisions to monster wand usage
The different mechanics for "monster like/don't like/can/can't use wand" were fairly inscrutable and full of special casing. Monsters were arbitrarily prevented from having wands beyond hd 9 even though a good number of wands were still capable of being relevant beyond that point if only the wand power scaled better.

Wand power was (30+hd) which is a pretty low range for power overall and meant that at early levels wands were devasting and at higher levels useless.

Monsters used a formula based on the number of charges players get per wand to decide whether wands were worthless past a certain point, which also left things error prone if anyone ever changes wand charges and doesn't realise that monster usage is connected (which has happened before).

So this commit makes rather more extensive changes than originally intended. Firstly there is a new "tier" system for monster wands to help decide at which HD ranges they are useful, and this now defines upper and lower limits. I removed the NO_HT_WANDS flag which was *only* used for Grinder and just made sure the new formula achieves pretty much the same thing.

Then the power scaling is given a much wider range and will scale a lot higher with hd, and additionally renamed the ALWAYS_WAND flag to ARTIFICER and gave monsters with this flag an additional power boost. This applies only to Ijyb and Maurice: the net effect is a slight nerf to Ijyb and a bit of a buff to Maurice. But the player is more likely to encounter random monsters who've picked up wands later in the dungeon and they'll know how to use them.

I moved most of the "can I use this" logic to monster::likes_wand and removed a number of redundant checks in the process; it's now impossible to roll digging for a monster because it's a separate list, and 0 charge wands should not have been possible to pick up anyway.

It should also be possible to give wands to more monsters in vaults now, if required.