Fix ^ screen for Jiyva and Ignis (#3851)
Apparently Webtiles uses regular expressions to parse the ^ screen,
and it really doesn't like the use of parentheses in strings
describing god powers. There is a way around this, but it puts the
resistance in the cost column on Webtiles (in the same way as
Cheibriados' stat boost), which is unacceptable.

Also change "injuring" to "harming" when referring to jellies,
for better flavour.