Adjust some Webtiles text colors
Magenta text in webtiles was conspicuously dark, making multiple status
indicators unnecessarily difficult to read against the game's black
background. (Lightmagenta was also extremely desaturated, and neither
looked particularly like their equivalents in local tiles).

This commit adjusts these to improve contrast and readability, and also
does this to a lesser degree to blues and dark red.

(Side note: not a great fan of how many things in this css seem to want to
reference a common color specifically, but only do so by copy-pasting it
with a name in comments. I've done my best to be careful and keep
everything in sync.)