Some refactoring and fixes for Aphotic Marionette
Instead of turning monster Good Neutral and using a temporary monster flag
to indicate that a monster should be considered a marionette while
spellcasting, this adds a new monster attitude (Marionette) that is
intended to be used only during the Marionette effect.

Note: The reason not to just make the monster Friendly during the effect
is that many spells have code to specifically prevent them being used by
friendly monsters, and this allows those spells to be cast (since the
player cannot be harmed by them anyway) while still firing tracers from the
perspective of 'hostile monsters are enemies and friendly ones are allies'.

This additionally fixes a few Marionette bugs:
 -Not getting XP for damage done by a marionette
 -Refusing to use spells properly near an invisible player if the
  marionette didn't have See Invis
 -In rare situations being able to still do harm to the player through
  non-damaging effects. (Marionettes just plain can fire through the player
  now, like shadows already could)