Buff Undying Armouries a little
They have an interesting set of abilities, but feel perhaps a little
undertuned for a rarer lategame enemy.

Make the dancing weapon created by Flashing Balestra spawn with Might and
increase the duration of Bestow Arms (and make ranged weapons a little more
likely to be given by it).

Also increase their spawn weight in the Vaults OOD table by a modest amount
(since it's the only place in the game they naturally spawn), trimming a
sliver of weight from storm and shadow dragons in the process.