Don't claim hellfire mortars 'flop about on dry land' (elliptic)
While I found it very hard to cause this in testing, it appears to relate
to being constricted in place for long enough that the lava path disappears
beneath it.

The main answer to this is to make the mortar too large to constrict, but
in the rare situation where it may be otherwise possible to replace the
lava beneath it with some other terrain, I have also suppressed the message
in general.