Make blocks of ice (and hoarfrost cannons) amphibious
The former in particular had some awkward interactions with being unable
to sculpt simulacra of merfolk while over deep water. The merfolk simulacra
would be able to swim, but the blocks of ice could not. Since blocks of ice
are pretty well known for floating in the ocean in reality, I don't think
it would be too weird if they did so here.

And let's let hoarfrost cannons do it too while we're at it, because why
not?