Remove -MP from Cloak of the Thief (elliptic)
The cloak having a downside at all, after its effect was changed, was
partially for historical reasons, and it seems plausible it doesn't need
one at all (it's already +0). Since -MP seems to still be stopping some of
the characters it could be interesting on from using it, let's try getting
rid of it entirely.