Let players choose which spell their Spellforged Servitor casts (elliptic)
Spellforged Servitor is a fun and popular spell, but it relies on a hidden
priority list of spells it can cast which isn't documented anywhere in
game. This is somewhat opaque to inexperienced players, and even for
experienced ones, can lead to unfortunate situations of wanting to avoid
memorizing a spell that the player would like to use, but which they would
rather their servitor not.

In an attempt to improve this situation a little, the player now has an
Imbue Servitor ability while Spellforged Servitor is known, which will
allow them to specify which eligable spell the player would like them to
cast.

This ability dismisses the servitor and takes 5 turns to complete, so is
not intended as a quick swap in combat. If the player never uses it, the
servitor defaults to its current behavior (to prevent people who forget or
do not realize from having a bad time). If the imbued spell somehow becomes
ineligable (either due to ?amnesia or the player no longer being able to
cast it well enough), the servitor will also fall back to the old default
behavior.

(It was suggested that it might be nice if the UI would also list spells in
the player's library which *could* be eligable if the player memorized them
or trained more, but that seems significantly more complicated to
implement. I wouldn't object if someone else tried, though.)