Fix a few breaks in mons_cast that should be returns
Breaks in the giant switch statement led to a default behavior of firing
the beam which is always passed into this function, after whatever the
spell's own effect was. In almost all cases, there should be a return
statement after the spell's code block instead (although at least two cases
do this break intentionally...)

Mostly, the effect of this seems to be invisible, though it caused
Marionette to crash when casting Summon Scorpions due to firing a beam with
no source position specified. I'm actually not sure why normal casting of
this spell *doesn't* crash, since setup_mons_cast() doesn't fill that info
in for the spell either (since it's not supposed to be *using* that beam),
but I will just assume something somewhere else along the way of the
normal cast path superfluously fills in this information.
f