Add a targeter to indicate which enemies will be misled by Shadowslip
I've seen a little uncertainty sometimes why some monsters were or were
not affected by this effect (possibly because the monsters weren't
targeting the player at the time), so now it hilights all such monsters
before cast.

I'm a bit concerned now that it might be misleading in a different way for
hilighting something different than 'where the player will move' but I hope
at least that part is unambiguous enough with experience to not require it?