Tweak Dithmenos wrath
Instead of summoning demons and undead to harass the player, Dithmenos can
now summon multiple shadow puppets (whose damage scales with the player's
level) with a chance to slow the player at the same time, or animate the
player's shadow against them as an active hostile monster.

The hostile player shadow can know a single shadow spell drawn at random
from the schools of spells the player knows. A handful of these don't
really work when cast by an independent shadow and so are excluded for the
moment.

I'd still like to do something with the Dith wrath option that puts you to
sleep, but this at least excises the 'arbitrarily evil' part of the wrath.