Make Beogh ally healing via damage stronger (and tweak wording)
I think the pseudo ally-vampirism effect Beogh granted was already decently
relevant, but seemed to have gone overlooked by a bunch of people. Since
Beogh doesn't seem in danger of being *too* good (and anything which
requires the player to put themselves in harm's way is already the 'less
strong' side of Beogh anyway), let's try buffing the effect a noticable
amount and see if people notice it more.