Fix pathfinding issues with apostle challenges / Blood for Blood
Pathfinding without a proper monster agent always considered closed doors
(and deep water and lava) to be untraversible, resulting in the code for
apostle challenges and BfB considering areas behind closed doors
unreachable. This could happen very frequently in Vaults.

Since apostle challenges that cannot be issued are merely delayed, and the
player will open doors as they explore the level, this may not have been
immediately obvious to most players, but it nonetheless shouldn't be
possible to avoid a challenge just by sitting in a room with the door
closed.

BfB had a similar issue due to how it places continuous reinforcements just
out of sight and has them move towards the player. It now should no longer
be flummoxed by doors.

(Friendly monsters normally cannot actually open doors, but I have made an
exception in this case, since it feels appropriate for the gameplay that
BfB creates.)