Fix Beogh ally-heal-on-damage not being active until 4*
When the UI claims 3*.

(I never seem to remember that piety_breakpoint(3) is actually 4* because
why the heck would it be? >.>)

(Note: it didn't even check your piety level at all until a couple days
ago, but it seemed better to have it match up with the UI's claims. It
wasn't really a balance adjustment, since it shows up as soon as you can
get apostles in the first place.)